rename component ImageSlider-> SliderItem
