+++ src/_styles/index.scss
@@ -10,4 +10,4 @@
 @import "socialMedia";
 @import "circularButton";
 @import "slider";
-@import "imageSlider";
+@import "sliderItem";
