+++ src/components/Slider.tsx
@@ -1,7 +1,7 @@
-import ImageSlider from "@components/ImageSlider";
+import CircularButton from "@components/CircularButton";
 import Container from "@components/layout/Container";
-import { FunctionComponent } from "react";
-
+import SliderItem from "@components/SliderItem";
+import { FunctionComponent, useState } from "react";
 const Slider: FunctionComponent = () => {
   const slides = [
     { url: "../img/project-1.gif", title: "Cards" },
@@ -11,10 +11,37 @@ const Slider: FunctionComponent = () => {
     { url: "../img/project-5.gif", title: "Game" },
   ];
 
+  const [currentIndex, setCurrentIndex] = useState(0);
+
+  const goToPrevious = (currentIndex) => {
+    const isFirstSlide = currentIndex === 0;
+    const newIndex = isFirstSlide ? slides.length - 1 : currentIndex - 1;
+    setCurrentIndex(newIndex);
+  };
+
+  const goToNext = (currentIndex) => {
+    const isLastSlide = currentIndex === slides.length - 1;
+    const newIndex = isLastSlide ? 0 : currentIndex + 1;
+    setCurrentIndex(newIndex);
+  };
+
   return (
     <Container>
-      <div className="slider-wrap">
-        <ImageSlider slides={slides} />
+      <div className="slider">
+        <CircularButton
+          direction="left"
+          handlerOnClick={() => goToPrevious(currentIndex)}
+          className="left-circular-button"
+        />
+        <SliderItem
+          slides={slides}
+          currentIndex={currentIndex}
+          setCurrentIndex={(newVal) => setCurrentIndex(newVal)}
+        />
+        <CircularButton
+          handlerOnClick={() => goToNext(currentIndex)}
+          className="right-circular-button"
+        />
       </div>
     </Container>
   );
